more refactoring - file and project reorganization
