+++ Example/Model/Scene.swift
@@ -1,18 +1,27 @@
 //
-//  TabViewController.swift
+//  Scene.swift
 //  Example
 //
 //  Created by Stefano Mondino on 23/07/2019.
 //  Copyright © 2019 Stefano Mondino. All rights reserved.
 //
 
+import Foundation
 import UIKit
 import PluginLayout
 
-enum DefaultScenes {
+enum Scene {
     case flow(pinned: Bool)
     case grid
     case staggered
+    
+    static var all: [Scene] {
+        return [
+            .flow(pinned: false),
+            .flow(pinned: true),
+            .staggered
+        ]
+    }
     
     var title: String {
         switch self {
@@ -27,7 +36,7 @@ enum DefaultScenes {
         case .flow (let pinned):
             let dataSource = DataSource(count: 40, contentType: .food, sections: 2)
             let delegate = FlowDelegate(dataSource: dataSource)
-//            let plugin = FlowLayoutPlugin(delegate: delegate, pinSectionHeaders: true, pinSectionFooters: true)
+            //            let plugin = FlowLayoutPlugin(delegate: delegate, pinSectionHeaders: true, pinSectionFooters: true)
             let layout = FlowLayout()
             layout.sectionFootersPinToVisibleBounds = pinned
             layout.sectionHeadersPinToVisibleBounds = pinned
@@ -35,25 +44,4 @@ enum DefaultScenes {
         default: return UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "simple") //?? UIViewController()
         }
     }
-}
-
-
-class TabViewController: UITabBarController {
-    
-    override func viewDidLoad() {
-    
-        super.viewDidLoad()
-        
-        self.viewControllers = [
-            DefaultScenes.flow(pinned: false),
-            DefaultScenes.flow(pinned: true),
-            .staggered
-            ].map {
-            let viewController = $0.viewController
-            viewController.title = $0.title
-            return UINavigationController(rootViewController: viewController)
-        }
-        
-    }
-    
 }
