+++ Example/ViewControllers/MixedViewController.swift
